ref(slack): Split assistant status internals
Separate assistant-status rendering, scheduling, and Slack sending so the status path is easier to reason about without changing the public runtime API.

Document the distinction between Slack-required behavior, Junior runtime policy, and product policy so future changes do not conflate transport requirements with UX decisions.

Refs GH-202
